Add files via upload
This is a camera rig with a few extra options that are usually available on other 3d applications and that I think can be of great help when doing your camera work. Besides some very practical parameters for camera animation it includes the option of viewing the camera cone and the DOF planes.